Working prototype
Basic prototype working on ripple emulator and phonegap >> iphone device (6S)

It needs some unit tests definitely at some point. And a way to make it distributable.

But for testing purposes there need to be multiple users. This is all directly on firebase user (hard coded) so that would need to change. For testing we can do it by entering user id in the settings. No security needed yet...

Question: do we want to use the built-in webSQL and synch to FB or is this for now okay?
Signed-off-by: ubilaz <erik@thirty-x.com>